Generate regular expressions from templates
