+++ .plop/regexps.hbs
@@ -1,0 +1,10 @@
+// START REGEXPS
+export const COLORREGS = {
+    [ColorModel.HEX]  : {{ HEX }},
+    [ColorModel.RGB]  : {{ RGB }},
+    [ColorModel.HSL]  : {{ HSL }},
+    [ColorModel.CMYK] : {{ CMYK }}
+};
+
+export const HSL_HUE = {{ HSL_HUE }};
+// END REGEXPS
