+++ package.json
@@ -25,6 +25,7 @@
     "build": "yarn clean && rollup --config rollup.config.js --bundleConfigAsCjs && yarn finish",
     "demo": "webpack serve --open --config webpack.demo.config.js",
     "docs": "webpack --config webpack.demo.publish.js --mode production",
+    "compile:regexps": "yarn plop",
     "prepare": "npm run build",
     "prepublishOnly": "npm run lint && npm run test",
     "version": "git add .",
@@ -71,6 +72,7 @@
     "jest": "^29.3.1",
     "mini-css-extract-plugin": "^2.7.2",
     "node-sass": "^8.0.0",
+    "plop": "^3.1.1",
     "rimraf": "^3.0.2",
     "rollup": "^3.8.1",
     "rollup-plugin-terser": "^7.0.2",
